theRadio/requirements_old.txt
absl-py
arrow
asn1crypto
astor
astroid
binaryornot
bleach
build-utils
certifi
cffi
chardet
Click
colorama
cookiecutter
coverage
cryptography
cycler
dataclasses
docutils
filelock
Flask
Flask-Login
Flask-PyMongo
future
gast
google-pasta
grpcio
h5py
helpdev
idna
importlib-metadata
inflect
irc
isort
itsdangerous
jaraco.classes
jaraco.collections
jaraco.functools
jaraco.itertools
jaraco.logging
jaraco.stream
jaraco.text
Jinja2
jinja2-git
jinja2-time
joblib
Keras
Keras-Applications
Keras-Preprocessing
kiwisolver
lazy-object-proxy
lxml
Markdown
MarkupSafe
matplotlib
mccabe
more-itertools
mypy
mypy-extensions
nose
ntlm-auth
panda
passlib
pkginfo
poyo
protobuf
pycparser
Pygments
pylint
pymongo
pyparsing
PyQt5
pyqt5plus
PyQt5-sip
python-bsonjs
python-dateutil
pytz
pywinrm
pyxDamerauLevenshtein
QDarkStyle
readme-renderer
redis
requests-ntlm
requests-toolbelt
scikit-learn
scipy
setuptools
six
sleekxmpp
sqlparse
s-tui
tempora
termcolor
Theano
tornado
tqdm
twine
typed-ast
typing-extensions
urllib3
urwid
webencodings
Werkzeug
wget
wheel
whichcraft
Whoosh
wrapt
xlrd
xmltodict
zipp
